requirements.txt
Flask==2.3.2
flask-cors==4.0.0
gunicorn==20.1.0

requests==2.28.2
structlog==22.1.0

google-auth==2.3.2

py-solc-x==1.1.1
web3==6.2.0
python-dotenv==1.0.0
marshmallow==3.19.0